Initial commit of tacplus-auth package
This package provides per-command TACACS authorization of commands.
It does that by being linked or symlinked to various commands in
a directory in the search path.  It uses the /etc/tacplus_servers
file to get the list of servers and other options.

In order to force a user to authorize commands, they need to be
in a restricted shell environment, with access to only the specific
commands that might be allowed.

The tacplus-restrict script can be used to setup the dot files
and directory permissions to keep the user restricted.